Totoro: aesthetic fixes and observability window intersection
- All (?) the imports have been modified to avoid relative imports.
In general, modules and not functions are imported. This should
prevent most of circular imports but dbclasses still need to be
imported lazily in many cases.

- getHARange, getLSTRange, and getUTRange in Plate and Set now accept
a keyword argument mjd used to define when the range should be
calculated. mjd='now' uses the current time. If mjd is not None,
the schedule file is used to determine the MaNGA observation window
and the returned intervals are intersected to return only the valid
observability windows.

git-svn-id: https://svn.sdss.org/repo/survey_software/sdss_python_module/trunk@4893 acfa968b-e9c4-4993-a9e6-846a59bbf143