fixed bug in plugger
Fixes a bug when a plugged plate cannot be observed for an integer
number of sets. In that case, the orphaned exposures are removed and the
remaining time is to be observed with a different plate. However, if there is
no other plate that can be used to observe at least one whole set, the
original, already plugged plate must be favoured.